Improved internal link loading
Now doesn’t need to wait for a post to be fetched before presenting the new view controller.
