+++ Client/Extensions/SwinjectStoryboardExtensions.swift
@@ -11,6 +11,7 @@ import SwinjectStoryboard
 extension SwinjectStoryboard {
     @objc class func setup() {
         let container = defaultContainer
+
         container.storyboardInitCompleted(NewsViewController.self) { resolver, controller in
             controller.authenticationUIService = resolver.resolve(AuthenticationUIService.self)!
             controller.swipeCellKitActions = resolver.resolve(SwipeCellKitActions.self)!
@@ -18,6 +19,7 @@ extension SwinjectStoryboard {
         container.storyboardInitCompleted(CommentsViewController.self) { resolver, controller in
             controller.authenticationUIService = resolver.resolve(AuthenticationUIService.self)!
             controller.swipeCellKitActions = resolver.resolve(SwipeCellKitActions.self)!
+            controller.navigationService = resolver.resolve(NavigationService.self)!
         }
         container.storyboardInitCompleted(SettingsViewController.self) { resolver, controller in
             controller.sessionService = resolver.resolve(SessionService.self)!
@@ -34,6 +36,9 @@ extension SwinjectStoryboard {
             SwipeCellKitActions(
                 authenticationUIService: resolver.resolve(AuthenticationUIService.self)!)
         }
+        container.register(NavigationService.self) { _ in
+            NavigationService()
+        }.inObjectScope(.container)
     }
 
     class func getService<T>() -> T? {
